+++ src/main/resources/application.properties
@@ -8,12 +8,15 @@ spring.datasource.username=postgres
 spring.datasource.password=rootmysql
 spring.datasource.driver-class-name=org.postgresql.Driver
 
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
+
+# ??? ?? ? validate? ??
+
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 
 jwt.secret=mySecureSecretKeyWithAtLeast32Characters
 jwt.expiration=3600000
 
-# PostgreSQL ???
+# PostgreSQL
 spring.jpa.properties.hibernate.jdbc.batch_size=500
 spring.jpa.properties.hibernate.order_inserts=true
